+++ src/ImageSprites/SpriteFragment.cs
@@ -3,8 +3,9 @@ namespace ImageSprites
 {
     internal class SpriteFragment
     {
-        public SpriteFragment(string fileName, int width, int height, int x, int y)
+        public SpriteFragment(string id, string fileName, int width, int height, int x, int y)
         {
+            ID = id;
             FileName = fileName;
             Width = width;
             Height = height;
@@ -12,8 +13,8 @@ namespace ImageSprites
             Y = y;
         }
 
+        public string ID { get; set; }
         public string FileName { get; set; }
-
         public int Width { get; set; }
         public int Height { get; set; }
         public int X { get; set; }