+++ src/ImageSprites/SpriteGenerator.cs
@@ -11,10 +11,10 @@ namespace ImageSprites
     {
         public async Task Generate(SpriteDocument doc)
         {
-            Dictionary<string, Bitmap> images = GetImages(doc);
+            var images = GetImages(doc);
 
-            int width = doc.Direction == Direction.Vertical ? images.Values.Max(i => i.Width) + (doc.Padding * 2) : images.Values.Sum(i => i.Width) + (doc.Padding * images.Count) + doc.Padding;
-            int height = doc.Direction == Direction.Vertical ? images.Values.Sum(img => img.Height) + (doc.Padding * images.Count) + doc.Padding : images.Values.Max(img => img.Height) + (doc.Padding * 2);
+            int width = doc.Direction == Direction.Vertical ? images.Values.Max(i => i.Item2.Width) + (doc.Padding * 2) : images.Values.Sum(i => i.Item2.Width) + (doc.Padding * images.Count) + doc.Padding;
+            int height = doc.Direction == Direction.Vertical ? images.Values.Sum(img => img.Item2.Height) + (doc.Padding * images.Count) + doc.Padding : images.Values.Max(img => img.Item2.Height) + (doc.Padding * 2);
 
             List<SpriteFragment> fragments = new List<SpriteFragment>();
 
@@ -30,7 +30,7 @@ namespace ImageSprites
                         Horizontal(images, fragments, canvas, doc.Padding);
                 }
 
-                string outputFile = Path.ChangeExtension(doc.FileName, doc.OutputExtension);
+                string outputFile = doc.FileName + doc.OutputExtension;
 
                 OnSaving(outputFile, doc);
                 bitmap.Save(outputFile, ImageHelpers.ExtensionFromFormat(doc.Format));
@@ -40,40 +40,45 @@ namespace ImageSprites
             await SpriteExporter.ExportStylesheet(fragments, doc, this);
         }
 
-        private static void Vertical(Dictionary<string, Bitmap> images, List<SpriteFragment> fragments, Graphics canvas, int margin)
+        private static void Vertical(Dictionary<string, Tuple<string, Bitmap>> images, List<SpriteFragment> fragments, Graphics canvas, int margin)
         {
             int currentY = margin;
 
-            foreach (string file in images.Keys)
+            foreach (string ident in images.Keys)
             {
-                var img = images[file];
-                fragments.Add(new SpriteFragment(file, img.Width, img.Height, margin, currentY));
+                var file = images[ident].Item1;
+                var img = images[ident].Item2;
+                fragments.Add(new SpriteFragment(ident, file, img.Width, img.Height, margin, currentY));
 
                 canvas.DrawImage(img, margin, currentY);
                 currentY += img.Height + margin;
             }
         }
 
-        private static void Horizontal(Dictionary<string, Bitmap> images, List<SpriteFragment> fragments, Graphics canvas, int margin)
+        private static void Horizontal(Dictionary<string, Tuple<string, Bitmap>> images, List<SpriteFragment> fragments, Graphics canvas, int margin)
         {
             int currentX = margin;
 
-            foreach (string file in images.Keys)
+            foreach (string ident in images.Keys)
             {
-                var img = images[file];
-                fragments.Add(new SpriteFragment(file, img.Width, img.Height, currentX, margin));
+                var file = images[ident].Item1;
+                var img = images[ident].Item2;
+                fragments.Add(new SpriteFragment(ident, file, img.Width, img.Height, currentX, margin));
 
                 canvas.DrawImage(img, currentX, margin);
                 currentX += img.Width + margin;
             }
         }
 
-        private static Dictionary<string, Bitmap> GetImages(SpriteDocument doc)
+        private static Dictionary<string, Tuple<string, Bitmap>> GetImages(SpriteDocument doc)
         {
-            Dictionary<string, Bitmap> images = new Dictionary<string, Bitmap>();
+            var images = new Dictionary<string, Tuple<string, Bitmap>>();
+            var source = doc.ToAbsoluteImages();
 
-            foreach (string file in doc.ToAbsoluteImages())
+            foreach (string ident in source.Keys)
             {
+                string file = source[ident];
+
                 if (!File.Exists(file))
                 {
                     throw new FileNotFoundException("One or more sprite input files don't exist", file);
@@ -82,7 +87,7 @@ namespace ImageSprites
                 var bitmap = (Bitmap)Image.FromFile(file);
                 bitmap.SetResolution(doc.Resolution, doc.Resolution);
 
-                images.Add(file, bitmap);
+                images.Add(ident, Tuple.Create(file, bitmap));
             }
 
             return images;